qtcreator: updated to 16.0.0
Qt Creator 16 comes with many performance improvements, especially for the
startup performance.

We improved the performance of the detection of the Qt ABI. Since Qt 6 we can
retrieve the ABI from a configuration file instead of guessing it from the
content of the libraries, which is now done asynchronously while collecting
other information about the registered Qt versions.