py-future: update to 1.0.0.
pkgsrc comment: this does not support Python 3.13.
The previous version didn't even support Python 3.12.

This package is still pulled in by ~15 packages, and if this
was a real problem, we probably would have noticed, so do not
mark this as not-for-3.13 for now.

Changes:

The new version number of 1.0.0 indicates that the python-future
project, like Python 2, is now done.

The most important change in this release is adding support for
Python 3.12 (ba1cc50 and a6222d2 and bcced95).

This release also includes these fixes:

    Small updates to the docs
    Add SECURITY.md describing security policy (0598d1b)
    Fix pasteurize: NameError: name ‘unicode’ is not defined (de68c10)
    Move CI to GitHub Actions (8cd11e8)
    Add setuptools to requirements for building docs (0c347ff)
    Fix typos in docs (350e87a)
    Make the fix_unpacking fixer more robust (de68c10)
    Small improvements to shell scripts according to shellcheck (6153844)