matchbox-panel: Improve detection of APM
Right now this package wrongly detects APM as being available on
NetBSD/sparc64 and NetBSD/alpha, leading to PLIST errors on those
platforms. It also fails to package on Linux platforms that don't
ship apm.h by default (which might be _most_ at this point).

Instead of assuming APM is available by default, only assume it's
available if the headers are in the expected places (which is what
the configure script checks anyway).

While this could be centralized in mk/platform, it's unlikely
to be widely useful considering how deprecated APM is as both
a hardware feature and API.