php*: Improve SUPERSEDES matches.
By using a generic catch-all of php-[0-9]* it was impossible for tools to
correctly determine what package should be the successor when there are
multiple to choose from.

Matching against exact versions will mean that improved SUPERSEDES handling
in the next version of pkgin should hopefully be able to support upgrades
Past the php renovation.  At least with a hand-modified pkgin.db and a test
pkgin binary it correctly detects the php-8.2.26 => php82-8.2.28nb1 rename.

Bump PKGREVISION for each.